little refactor on the x-ray (personal) section
