+++ src/app/page.module.css
@@ -1,12 +1,6 @@
 .main {
 }
 
-.main > section:nth-child(1) {
-  background-image: url("/personal/mountains.png");
-  background-repeat: no-repeat;
-  background-size: cover;
-}
-
 .main > section {
   /* TODO Please remove this */
   min-height: 450px;
